Find known_hosts entries when HashKnownHosts=yes
When HashKnownHosts=yes, known_hosts entries are no longer necessarily
of the form: `host[,host]* type key`

Lines can also be in the form: `|1|salt|hmac type key`

The '|1|' indicates the method used to hash the host, but currently only
'1' exists, and it specifies the use of a SHA1-based HMAC.

'salt' and 'hmac' are both Base64-encoded.  'salt' is the key to use
with the HMAC, and 'hmac' is the outcome.